Move health factor panel content
