+++ packages/app/src/features/easy-borrow/components/side-panel/EasyBorrowSidePanel.tsx
@@ -1,7 +1,7 @@
 import { formatPercentage } from '@/domain/common/format'
 import { LiquidationDetails } from '@/domain/market-info/getLiquidationDetails'
 import { Panel } from '@/ui/atoms/panel/Panel'
-import { HealthFactorPanelContent } from '@/ui/molecules/new/health-factor-panel-content/HealthFactorPanelContent'
+import { HealthFactorPanelContent } from '@/ui/molecules/health-factor-panel-content/HealthFactorPanelContent'
 import { testIds } from '@/ui/utils/testIds'
 import BigNumber from 'bignumber.js'
 import { BorrowDetails } from '../../logic/useEasyBorrow'
